New rule 'cypress-no-force'
The Cypress rule 'no-force' only works when calling the function like
`cy.get(...).click({ force: true })`, i.e. with `cy` in the same line.
But in our code we often use helper functions, so the rule does not
detect our usages. This new version of the rule leverages TypeScript to
work out if the method is called on a `Cypress.Chainable` object.