Make `recommended` generic and create `recommended-angular`
When we first created the plugin, we only really thought of Angular apps
at Criteo. Later, new configs were added for React, but `recommended`,
`recommended-app`, `recommended-lib` and `recommended-template` remained
Angular-specific. This commit aims to change that by creating new config
files with Angular in the name. To avoid breaking changes,
`recommended-app`, `recommended-lib` and `recommended-template` are still
available, but are deprecated in favour of the Angular-specific names.