Reorganise files so that "extends" works properly
The relative paths in "extends" are treated differently depending if
it's at the top level (directly extended from the client's .eslintrc) or
extended from another config file internally. To get around this, I have
created a directory "internal" where the bulk of the work is done, and
then a public layer which extends from the files in "internal".

Concretely, this allows us to expose the "recommended" config, while
also extending it for other configs such as "recommended-angular".

I also moved some rules to "recommended", per Xavier's comments.